# docs/swagger.yaml
basePath: /
definitions:
...
  models.Train:
    properties:
      company:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  models.Trip:
    properties:
      created_at:
        type: string
      direction:
        type: string
      end_time:
        type: string
      id:
        type: integer
      start_time:
        type: string
      train:
        allOf:
        - $ref: '#/definitions/models.Train'
        description: Foreign key relationships
      train_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A Metro Card Management System API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Metro API
  version: "1.0"
paths:
...
  /trip:
    get:
      consumes:
      - application/json
      description: Retrieve all trip records with optional filtering
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of records per page
        in: query
        name: limit
        type: integer
      - description: Filter by train ID
        in: query
        name: train_id
        type: integer
      - description: Filter by direction
        in: query
        name: direction
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Trips retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Trip'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get all trip records
      tags:
      - trip
    post:
      consumes:
      - application/json
      description: Create a new trip record for train journeys
      parameters:
      - description: Trip information
        in: body
        name: trip
        required: true
        schema:
          $ref: '#/definitions/models.Trip'
      produces:
      - application/json
      responses:
        "200":
          description: Trip created successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Trip'
              type: object
        "400":
          description: Bad request - validation error
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Create a new trip record
      tags:
      - trip
  /trip/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a trip record
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Trip deleted successfully
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Trip not found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Delete trip
      tags:
      - trip
    get:
      consumes:
      - application/json
      description: Retrieve a specific trip record by its ID
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Trip retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Trip'
              type: object
        "404":
          description: Trip not found
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get trip by ID
      tags:
      - trip
    put:
      consumes:
      - application/json
      description: Update an existing trip record
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated trip information
        in: body
        name: trip
        required: true
        schema:
          $ref: '#/definitions/models.Trip'
      produces:
      - application/json
      responses:
        "200":
          description: Trip updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Trip'
              type: object
        "400":
          description: Bad request - validation error
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Trip not found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Update trip
      tags:
      - trip
  /trip/active:
    get:
      consumes:
      - application/json
      description: Retrieve all currently active trips (where end_time is null or
        in the future)
      produces:
      - application/json
      responses:
        "200":
          description: Active trips retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Trip'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get active trips
      tags:
      - trip
  /trip/direction/{direction}:
    get:
      consumes:
      - application/json
      description: Retrieve all trip records for a specific direction
      parameters:
      - description: Direction
        in: path
        name: direction
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Trips retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Trip'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get trips by direction
      tags:
      - trip
  /trip/train/{train_id}:
    get:
      consumes:
      - application/json
      description: Retrieve all trip records for a specific train
      parameters:
      - description: Train ID
        in: path
        name: train_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Trips retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Trip'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get trips by train ID
      tags:
      - trip
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"